requirements.txt
tesseract==0.1.3
pillow==11.3.0
cvs==1.12.13
opencv-python==4.11.0.86
pandas==2.3.0